requirements.txt
pymilvus>=2.3.0
sentence-transformers>=2.2.0
beautifulsoup4>=4.12.0
requests>=2.31.0
openai>=1.0.0
python-dotenv>=1.0.0
html5lib>=1.1
streamlit>=1.50.0
